added filterit as an internal qmail-local command
