+++ indimail-mta-x/filterit.h
@@ -1,0 +1,16 @@
+/*
+ * $Id: $
+ */
+#ifndef _FILTERIT_H_
+#define _FILTERIT_H_
+
+#define FATAL "filterit: fatal: "
+#define WARN  "filterit: warn: "
+
+int             filterit_sub1(int, char **);
+int             filterit_sub2(char *);
+#endif
+
+/*
+ * $Log: $
+ */
